Module 2.2 EDA and Feature Engineering
you will see if the data can be used to automatically determine if the Falcon 9’s first stage will land. Some attributes can be used to determine if the first stage can be reused. We can then use these features with machine learning to automatically predict if the first stage can land successfully