Module 3.1 Interactive Visual Analytics Folium
The first part of this module will be focused on analyzing launch site geo and proximities with Folium. We will first mark the launch site locations and their close proximities on an interactive map. Then, we can explore the map with those markers and try to discover any patterns from them. Finally, we should be able to explain how to choose an optimal launch site.