Create Module 3.2 Interactive Dashboard with Ploty Dash
you will be building a dashboard application with the Python Plotly Dash package. This dashboard application contains input components such as a dropdown list and a range slider to interact with a pie chart and a scatter point chart. You will be guided to build this dashboard application in an instructional lab. After the dashboard is built, you can use it to find more insights from the SpaceX dataset more easily than with static graphs.